+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 -   `cy.solidAuthorize()` command.
+-   Javascript evaluation in replacement templates, with `now()`, `date('2024-03-03')`, and `uuid()` helpers.
 
 ## [v0.1.3](https://github.com/NoelDeMartin/cypress-solid/releases/tag/v0.1.3) - 2024-01-13