+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [v0.2.0](https://github.com/NoelDeMartin/cypress-solid/releases/tag/v0.2.0) - 2025-05-25
 
 ### Added
 
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Modernized tooling.
+- `solidLogin` and `solidAuthorize` commands now wait until going back to the app before resolving.
 
 ## [v0.1.4](https://github.com/NoelDeMartin/cypress-solid/releases/tag/v0.1.4) - 2024-05-10